webrtc_ice::agent::agent_config

Struct AgentConfig

pub struct AgentConfig {

    pub urls: Vec<Url>,
    pub udp_network: UDPNetwork,
    pub local_ufrag: String,
    pub local_pwd: String,
    pub multicast_dns_mode: MulticastDnsMode,
    pub multicast_dns_host_name: String,
    pub multicast_dns_dest_addr: String,
    pub disconnected_timeout: Option<Duration>,
    pub failed_timeout: Option<Duration>,
    pub keepalive_interval: Option<Duration>,
    pub network_types: Vec<NetworkType>,
    pub candidate_types: Vec<CandidateType>,
    pub check_interval: Duration,
    pub max_binding_requests: Option<u16>,
    pub is_controlling: bool,
    pub lite: bool,
    pub nat_1to1_ip_candidate_type: CandidateType,
    pub nat_1to1_ips: Vec<String>,
    pub host_acceptance_min_wait: Option<Duration>,
    pub srflx_acceptance_min_wait: Option<Duration>,
    pub prflx_acceptance_min_wait: Option<Duration>,
    pub relay_acceptance_min_wait: Option<Duration>,
    pub net: Option<Arc<Net>>,
    pub interface_filter: Arc<Option<InterfaceFilterFn>>,
    pub ip_filter: Arc<Option<IpFilterFn>>,
    pub insecure_skip_verify: bool,
    pub include_loopback: bool,

}

Collects the arguments to ice::Agent construction into a single structure, for future-proofness of the interface.

Fields
urls: Vec<Url>udp_network: UDPNetwork

Controls how the UDP network stack works. See UDPNetwork

local_ufrag: String

It is used to perform connectivity checks. The values MUST be unguessable, with at least 128 bits of random number generator output used to generate the password, and at least 24 bits of output to generate the username fragment.

local_pwd: String

It is used to perform connectivity checks. The values MUST be unguessable, with at least 128 bits of random number generator output used to generate the password, and at least 24 bits of output to generate the username fragment.

multicast_dns_mode: MulticastDnsMode

Controls mDNS behavior for the ICE agent.

multicast_dns_host_name: String

Controls the hostname for this agent. If none is specified a random one will be generated.

multicast_dns_dest_addr: String

Control mDNS destination address

disconnected_timeout: Option<Duration>

Defaults to 5 seconds when this property is nil. If the duration is 0, the ICE Agent will never go to disconnected.

failed_timeout: Option<Duration>

Defaults to 25 seconds when this property is nil. If the duration is 0, we will never go to failed.

keepalive_interval: Option<Duration>

Determines how often should we send ICE keepalives (should be less then connectiontimeout above) when this is nil, it defaults to 10 seconds. A keepalive interval of 0 means we never send keepalive packets

network_types: Vec<NetworkType>

An optional configuration for disabling or enabling support for specific network types.

candidate_types: Vec<CandidateType>

An optional configuration for disabling or enabling support for specific candidate types.

check_interval: Duration

Controls how often our internal task loop runs when in the connecting state. Only useful for testing.

max_binding_requests: Option<u16>

The max amount of binding requests the agent will send over a candidate pair for validation or nomination, if after max_binding_requests the candidate is yet to answer a binding request or a nomination we set the pair as failed.

is_controlling: boollite: bool

lite agents do not perform connectivity check and only provide host candidates.

nat_1to1_ip_candidate_type: CandidateType

It is used along with nat1to1ips to specify which candidate type the 1:1 NAT IP addresses should be mapped to. If unspecified or CandidateTypeHost, nat1to1ips are used to replace host candidate IPs. If CandidateTypeServerReflexive, it will insert a srflx candidate (as if it was derived from a STUN server) with its port number being the one for the actual host candidate. Other values will result in an error.

nat_1to1_ips: Vec<String>

Contains a list of public IP addresses that are to be used as a host candidate or srflx candidate. This is used typically for servers that are behind 1:1 D-NAT (e.g. AWS EC2 instances) and to eliminate the need of server reflexisive candidate gathering.

host_acceptance_min_wait: Option<Duration>

Specify a minimum wait time before selecting host candidates.

srflx_acceptance_min_wait: Option<Duration>

Specify a minimum wait time before selecting srflx candidates.

prflx_acceptance_min_wait: Option<Duration>

Specify a minimum wait time before selecting prflx candidates.

relay_acceptance_min_wait: Option<Duration>

Specify a minimum wait time before selecting relay candidates.

net: Option<Arc<Net>>

Net is the our abstracted network interface for internal development purpose only (see (github.com/pion/transport/vnet)[github.com/pion/transport/vnet]).

interface_filter: Arc<Option<InterfaceFilterFn>>

A function that you can use in order to whitelist or blacklist the interfaces which are used to gather ICE candidates.

ip_filter: Arc<Option<IpFilterFn>>

A function that you can use in order to whitelist or blacklist the ips which are used to gather ICE candidates.

insecure_skip_verify: bool

Controls if self-signed certificates are accepted when connecting to TURN servers via TLS or DTLS.

include_loopback: bool

Include loopback addresses in the candidate list.