webrtc_ice

Enum Error

#[non_exhaustive]
pub enum Error {

    ErrUnknownType,
    ErrSchemeType,
    ErrStunQuery,
    ErrInvalidQuery,
    ErrHost,
    ErrPort,
    ErrLocalUfragInsufficientBits,
    ErrLocalPwdInsufficientBits,
    ErrProtoType,
    ErrClosed,
    ErrNoCandidatePairs,
    ErrCanceledByCaller,
    ErrMultipleStart,
    ErrRemoteUfragEmpty,
    ErrRemotePwdEmpty,
    ErrNoOnCandidateHandler,
    ErrMultipleGatherAttempted,
    ErrUsernameEmpty,
    ErrPasswordEmpty,
    ErrAddressParseFailed,
    ErrLiteUsingNonHostCandidates,
    ErrUselessUrlsProvided,
    ErrUnsupportedNat1to1IpCandidateType,
    ErrInvalidNat1to1IpMapping,
    ErrExternalMappedIpNotFound,
    ErrMulticastDnsWithNat1to1IpMapping,
    ErrIneffectiveNat1to1IpMappingHost,
    ErrIneffectiveNat1to1IpMappingSrflx,
    ErrInvalidMulticastDnshostName,
    ErrRestartWhenGathering,
    ErrRunCanceled,
    ErrTcpMuxNotInitialized,
    ErrTcpRemoteAddrAlreadyExists,
    ErrSendPacket,
    ErrAttributeTooShortIceCandidate,
    ErrParseComponent,
    ErrParsePriority,
    ErrParsePort,
    ErrParseRelatedAddr,
    ErrParseType,
    ErrUnknownCandidateType,
    ErrGetXorMappedAddrResponse,
    ErrConnectionAddrAlreadyExist,
    ErrReadingStreamingPacket,
    ErrWriting,
    ErrClosingConnection,
    ErrDetermineNetworkType,
    ErrMissingProtocolScheme,
    ErrTooManyColonsAddr,
    ErrRead,
    ErrUnknownRole,
    ErrMismatchUsername,
    ErrIceWriteStunMessage,
    ErrInvalidUrl,
    ErrUrlParse,
    ErrCandidateIpNotFound,
    ParseInt(ParseIntError),
    ParseIp(AddrParseError),
    Io(IoError),
    Util(Error),
    Stun(Error),
    ParseUrl(ParseError),
    Mdns(Error),
    Turn(Error),
    Other(String),

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ErrUnknownType

Indicates an error with Unknown info.

ErrSchemeType

Indicates the scheme type could not be parsed.

ErrStunQuery

Indicates query arguments are provided in a STUN URL.

ErrInvalidQuery

Indicates an malformed query is provided.

ErrHost

Indicates malformed hostname is provided.

ErrPort

Indicates malformed port is provided.

ErrLocalUfragInsufficientBits

Indicates local username fragment insufficient bits are provided. Have to be at least 24 bits long.

ErrLocalPwdInsufficientBits

Indicates local passoword insufficient bits are provided. Have to be at least 128 bits long.

ErrProtoType

Indicates an unsupported transport type was provided.

ErrClosed

Indicates the agent is closed.

ErrNoCandidatePairs

Indicates agent does not have a valid candidate pair.

ErrCanceledByCaller

Indicates agent connection was canceled by the caller.

ErrMultipleStart

Indicates agent was started twice.

ErrRemoteUfragEmpty

Indicates agent was started with an empty remote ufrag.

ErrRemotePwdEmpty

Indicates agent was started with an empty remote pwd.

ErrNoOnCandidateHandler

Indicates agent was started without on_candidate.

ErrMultipleGatherAttempted

Indicates GatherCandidates has been called multiple times.

ErrUsernameEmpty

Indicates agent was give TURN URL with an empty Username.

ErrPasswordEmpty

Indicates agent was give TURN URL with an empty Password.

ErrAddressParseFailed

Indicates we were unable to parse a candidate address.

ErrLiteUsingNonHostCandidates

Indicates that non host candidates were selected for a lite agent.

ErrUselessUrlsProvided

Indicates that one or more URL was provided to the agent but no host candidate required them.

ErrUnsupportedNat1to1IpCandidateType

Indicates that the specified NAT1To1IPCandidateType is unsupported.

ErrInvalidNat1to1IpMapping

Indicates that the given 1:1 NAT IP mapping is invalid.

ErrExternalMappedIpNotFound

IPNotFound in NAT1To1IPMapping.

ErrMulticastDnsWithNat1to1IpMapping

Indicates that the mDNS gathering cannot be used along with 1:1 NAT IP mapping for host candidate.

ErrIneffectiveNat1to1IpMappingHost

Indicates that 1:1 NAT IP mapping for host candidate is requested, but the host candidate type is disabled.

ErrIneffectiveNat1to1IpMappingSrflx

Indicates that 1:1 NAT IP mapping for srflx candidate is requested, but the srflx candidate type is disabled.

ErrInvalidMulticastDnshostName

Indicates an invalid MulticastDNSHostName.

ErrRestartWhenGathering

Indicates Restart was called when Agent is in GatheringStateGathering.

ErrRunCanceled

Indicates a run operation was canceled by its individual done.

ErrTcpMuxNotInitialized

Initialized Indicates TCPMux is not initialized and that invalidTCPMux is used.

ErrTcpRemoteAddrAlreadyExists

Indicates we already have the connection with same remote addr.

ErrSendPacket

ErrAttributeTooShortIceCandidate

ErrParseComponent

ErrParsePriority

ErrParsePort

ErrParseRelatedAddr

ErrParseType

ErrUnknownCandidateType

ErrGetXorMappedAddrResponse

ErrConnectionAddrAlreadyExist

ErrReadingStreamingPacket

ErrWriting

ErrClosingConnection

ErrDetermineNetworkType

ErrMissingProtocolScheme

ErrTooManyColonsAddr

ErrRead

ErrUnknownRole

ErrMismatchUsername

ErrIceWriteStunMessage

ErrInvalidUrl

ErrUrlParse

ErrCandidateIpNotFound

ParseInt(ParseIntError)

ParseIp(AddrParseError)

Io(IoError)

Util(Error)

Stun(Error)

ParseUrl(ParseError)

Mdns(Error)

Turn(Error)

Other(String)