webrtc_ice::udp_mux

Struct UDPMuxConn

pub struct UDPMuxConn { /* private fields */ }

A UDP mux connection.

Implementations

impl UDPMuxConn

pub fn new(params: UDPMuxConnParams) -> Self

Creates a new UDPMuxConn.

pub fn key(&self) -> &str

Returns a key identifying this connection.

pub async fn write_packet( &self, data: &[u8], addr: SocketAddr, ) -> Result<(), Error>

Writes data to the given address. Returns an error if the buffer is too short or there’s an encoding error.

pub fn is_closed(&self) -> bool

Returns true if this connection is closed.

pub fn close_rx(&self) -> Receiver<bool>

Gets a copy of the close tokio::sync::watch::Receiver that fires when this connection is closed.

pub fn close(&self)

Closes this connection.

pub fn get_addresses(&self) -> Vec<SocketAddr>

Gets the list of the addresses associated with this connection.

pub async fn add_address(&self, addr: SocketAddr)

Registers a new address for this connection.

pub fn remove_address(&self, addr: &SocketAddr)

Deregisters an address.

pub fn contains_address(&self, addr: &SocketAddr) -> bool

Returns true if the given address is associated with this connection.