webrtc_ice::udp_mux

Trait UDPMuxWriter

pub trait UDPMuxWriter {
    // Required methods
    fn register_conn_for_address<'life0, 'life1, 'async_trait>(
        &'life0 self,
        conn: &'life1 UDPMuxConn,
        addr: SocketAddr,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn send_to<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        buf: &'life1 [u8],
        target: &'life2 SocketAddr,
    ) -> Pin<Box<dyn Future<Output = Result<usize, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
}

A trait for a UDPMuxConn to communicate with an UDP mux.

Required Methods

fn register_conn_for_address<'life0, 'life1, 'async_trait>( &'life0 self, conn: &'life1 UDPMuxConn, addr: SocketAddr, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Registers an address for the given connection.

fn send_to<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, buf: &'life1 [u8], target: &'life2 SocketAddr, ) -> Pin<Box<dyn Future<Output = Result<usize, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Sends the content of the buffer to the given target.

Returns the number of bytes sent or an error, if any.